renamed loggerfunc into errorlogger for clarity
